Migrate queries from live specs api file
